aider: Add __init__.py files to make directories into Python packages.
# Aider chat conversation:

USER: in order to make the modules discovarable should i add the file __init__.py in each subdirectory of the project?

ASSISTANT: Yes, adding an `__init__.py` file in each subdirectory of your project will make these directories into Python packages, and thus their modules will be discoverable by Python when you try to import them. This is a common practice in Python projects to signal to Python that a directory should be treated as a package.

Would you like me to generate the diffs for adding `__init__.py` files to each subdirectory? If so, please provide the list of subdirectories that should be turned into packages.

USER: convert the `postgres_da_ai_agent' and all subdirectories into packages

ASSISTANT: To convert the `postgres_da_ai_agent` directory and all its subdirectories into Python packages, we will add an `__init__.py` file to each of them. Here are the diffs for creating these files:

```diff
--- /dev/null
+++ postgres_da_ai_agent/__init__.py
@@ -0,0 +1 @@
+__all__ = ["agents", "modules", "prompt_handler"]
```

```diff
--- /dev/null
+++ postgres_da_ai_agent/agents/__init__.py
@@ -0,0 +1 @@
+__all__ = ["instruments", "turbo4"]
```

```diff
--- /dev/null
+++ postgres_da_ai_agent/modules/__init__.py
@@ -0,0 +1 @@
+__all__ = ["db", "embeddings", "llm", "rand"]
```

These diffs represent the creation of new `__init__.py` files in each of the specified directories. The `__all__` list is optional and can be used to explicitly specify which modules the package exports as the API. If you don't want to restrict the exported modules, you can simply create an empty `__init__.py` file.